+++ assets/js/content/consumable.js
@@ -59,6 +59,17 @@ window.Consumable = (function() {
 					"sound": "cautious",
 				},
 			},
+			"cloud": {
+				"brown": {
+					"name": "brown",
+					"type": "cloud",
+					"ref": "a",
+					"img": "brown-cloud.gif",
+					"val": -FS_Number.round(Math.random() * 0.2, 2),
+					"stat": "health",
+					"sound": "cautious",
+				},
+			},
 		};
 
 	/**
@@ -85,10 +96,11 @@ window.Consumable = (function() {
 			// pick random from type
 			if (r < 0.05) create("cookie", "", count);
 			else if (r < 0.06) create("junk", "", count);
+			else if (r < 0.07) create("cloud", "", count);
 			// pick random type
-			else if (r < 0.07) create("", "", count);
+			else if (r < 0.08) create("", "", count);
 			// gameMode === testing
-			else if (r < 0.4 && ["demo","testing"].includes(tally_options.gameMode)) create("cookie", "", 1);
+			else if (r < 0.4 && ["demo","testing"].includes(tally_options.gameMode)) create("cloud", "", 1);
 
 		} catch (err) {
 			console.error(err);
@@ -121,6 +133,7 @@ window.Consumable = (function() {
 
 				// testing
 				//consumables.push(types.cookie.fortune);
+				// consumables.push(types.cloud.brown);
 			}
 			if (DEBUG) console.log("🍪 Consumable.create()", consumables);
 			add();
@@ -152,8 +165,13 @@ window.Consumable = (function() {
 				// html
 				imgStr = chrome.extension.getURL('assets/img/consumables/' + consumables[i].type + "/" + consumables[i].img);
 				id = 'tally_consumable_' + i;
-				str = "<div data-consumable='" + i + "' class='tally_consumable_inner' id='" + id + "' style='" + css + "'>" +
-					"<img src='" + imgStr + "'></div>";
+				str = "<div data-consumable='" + i + "' class='tally_consumable_inner' id='" + id + "' style='" + css + "'>";
+				str += "<img src='" + imgStr + "'";
+				// make clouds semi-transparent
+				if (consumables[i].type == "cloud"){
+					str += " style='opacity:.7';";
+				}
+				str += "></div>";
 				$('.tally_consumable_wrapper').append(str);
 
 				// add listeners